+++ code/system.ui/react.ui.monaco/src/test.ui/entry.Specs.mts
@@ -1,5 +1,5 @@
 export const Specs = {
-  'sys.ui.code.MonacoEditor': () => import('../ui/MonacoEditor/-SPEC'),
+  'sys.ui.code.MonacoEditor': () => import('../ui/MonacoEditor/-dev/-SPEC'),
   'sys.ui.code.MonacoCrdt': () => import('../ui.logic/MonacoCrdt/-dev/-SPEC'),
 };
